+++ pkg/cli/cmd/bulkexport/iamresource/iamresource.go
@@ -17,7 +17,7 @@ package iamresource
 import (
 	"fmt"
 
-	"github.com/GoogleCloudPlatform/k8s-config-connector/pkg/apis/iam/v1beta1"
+	"github.com/GoogleCloudPlatform/k8s-config-connector/apis/iam/v1beta1"
 	"github.com/GoogleCloudPlatform/k8s-config-connector/pkg/deepcopy"
 
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
